pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sap-process-mining"
version = "0.1.0"
description = "AI-powered SAP process mining — continuous operational intelligence from SAP event logs to plain-language insight"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Kiran", email = "kiran.banavara@gmail.com" }]
license = { text = "Proprietary" }
dependencies = [
    "pandas>=2.1",
    "pyarrow>=15.0",           # parquet cache of event logs for the Investigator
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "python-dateutil>=2.8",
    "python-dotenv>=1.0",      # auto-load .env at CLI / UI entry
    "httpx>=0.27",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
gemini = ["google-genai>=0.3"]
ui = ["streamlit>=1.36"]
hana = ["hdbcli>=2.20"]
rfc = ["pyrfc>=3.3"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.6"]
all = [
    "anthropic>=0.40",
    "openai>=1.50",
    "google-genai>=0.3",
    "streamlit>=1.36",
]

[project.scripts]
sap-mining = "sap_process_mining.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
